rename(UserEntity): Change import paths of domain models
